refactor: split hardhat utils into core/bridge folders
