+++ typescript/optics-tests/lib/bridge.ts
@@ -1,0 +1,8 @@
+export enum BridgeMessageTypes {
+  INVALID = 0,
+  TOKEN_ID,
+  MESSAGE,
+  TRANSFER,
+  DETAILS,
+  REQUEST_DETAILS,
+}
