+++ typescript/optics-tests/test/cross-chain/recoveryManager.test.ts
@@ -5,7 +5,7 @@ import * as types from 'ethers';
 import { formatCall, sendFromSigner } from './utils';
 import { increaseTimestampBy } from '../utils';
 import { getTestDeploy } from '../testChain';
-import { Updater } from '../../lib';
+import { Updater } from '../../lib/core';
 import { Signer } from '../../lib/types';
 import { CoreDeploy as Deploy } from '../../../optics-deploy/src/core/CoreDeploy';
 import { deployTwoChains } from '../../../optics-deploy/src/core';
