+++ typescript/optics-tests/test/home.test.ts
@@ -1,7 +1,7 @@
 import { ethers, optics } from 'hardhat';
 import { expect } from 'chai';
 import { getTestDeploy } from './testChain';
-import { OpticsState, Updater } from '../lib';
+import { OpticsState, Updater } from '../lib/core';
 import { Signer } from '../lib/types';
 import { CoreDeploy as Deploy } from '../../optics-deploy/src/core/CoreDeploy';
 import * as deploys from '../../optics-deploy/src/core';
